move IsEmptyDir to configs package
